MVC implementation of sticky versions in versions dropdown
We check the CI_COMMIT_REF_NAME when the site is built and apply it
to the versions dropdown.

Example:

For 10.4, the entire site will have 10.4 always displayed in the
dropdown, except for the archives page which should always point to the
most recent archives and thus its name is just 'Versions' (we don't care
about the version).