Move archives content to an index.html file
This fixes the issue when on the /archives page.

The dropdown login is that it takes the Nanoc identifier (file path
under content/) and in the archives case it was different than the
created file (set in Rules).

With this change, the identifier and the ending result are the same,
thus the dropdown link no longer 404.