Add environments to the Registry jobs
Sometimes, we need to check on the jobs that build the
single versions, but since they are run once a month,
it's almost impossible to find them (no search in pipelines).

This adds a registry/ environment to all jobs that upload
Docker images to the Container Registry, so that we can
easily browse them through the "Operations > Environments" page.