Symlink `/ce/` to `/ee/`.
Now that we have a single codebase, the CE and EE docs
are identical, thus duplicate. This leads to having almost
twice the size of artifacts, which can cause Pages errors.

With this change, we remove the CE HTML content and use a symlink to EE
instead. The structure will remain the same without the duplicate
content.

Let GitLab CI handle the symlinks.