Make sure we also capture full URLs when converting the versioned URLs
In cross-product docs, we're referencing full URLs instead of relative
ones. For example, an Omnibus doc can have a doc reference to a Runner
doc URL. In that case, we use the full URL.

This change adds another find command that checks and replaces the
version for full URLs as well.