Print the last commit of a branch
This is useful if we want to confirm that the correct changes were
pulled from the upstream project. When we build the review apps
multiple times during a review process, sometimes the content
might not update. Printing the last commit that was pulled will
help us compare to the last commit of the upstream project.