Move Monitor docs to separate top-level menu
The Monitor docs have been separated out across multiple places,
and not organized in a way that made the stage happy. This MR takes
the existing documentation for Monitor and moves it into a new
top-level menu item, "Operations."

Many items are missing, because they are all jammed into the giant
(7000-word) Prometheus page. As pages are split out of the
Prometheus page, slimming it down, they'll get added to the
Operations menu.