Check if the product is present and hide the topics if not
We had switched to relative in the past
https://gitlab.com/gitlab-org/gitlab-docs/-/merge_requests/734, but we
then used full URLs again
https://gitlab.com/gitlab-org/gitlab-docs/-/merge_requests/742.

The reason is that we use this project as a Docker image
(`registry.gitlab.com/gitlab-org/gitlab-docs:docs-lint`) to build and
test other projects like Omnibus, Runner, etc. When doing so, the
GitLab project is not present, so the `ee/` relative paths
(and `omnibus/` respectively) will fail the URL test we have in place.

With this change, we first check if the content is present and
show the topics if so. We hide them otherwise, so that the
internal_links test won't fail.