Encode Mermaid HTML to htmlsafe values
There's an inconsistency with some filters, for example
admonition and introduced_in, that call Nokogiri::HTML#fragment
on the input and write the output, which means code that looks like
tags gets treated as actual HTML tags.

This breaks mermaid diagrams because those tags are not valid and
the Mermaid JS parser trips over them, converting them to the
html-encoded values is what was currently happening for ->> strings
and it's working, so pre-converting all of them gets rid of the problem.