Allow alerts to stand on their own
Previously, for the alerts to work, a whitespace was needed to succeed
them, and this was possible since we were using `NOTE: **Note:**`.

However, if we wanted to get rid of the `**Note:**` text, this would mean
that the alert would need a trailing whitespace in order to be rendered
and it would look like `NOTE: `.

With that change, the regex is more permissive and allows the whitespace
to be absent.

Since now we can render alerts with only the icon present, an alternate
title was also added for accessibility issues.