Hide survey banner on stable branches
We don't want the banner to be shown on stable branches.
This introduces a new function that checks if CI_COMMIT_REF_NAME
is defined, and if not, it uses Git to find the current branch name.

If show_banner is false and current_branch is not a stable branch,
hide the banner.