Do not render the home page in upstream products
The home page may contain relative links to one
of the four products we build the docs from that
may not be present when the site is built and
tested in the upstream products.

For example, when the docs site is built and
tested for dead relative links in the Runner
project, it will fail for some relative links
in the front page that point to the Omnibus
docs, since the Omnibus docs are absent.