Skip feature flags generation if yaml is absent
Feature flags are generated dynamically when
`content/_data/feature_flags.yaml` is present.

For upstream projects where this file is not
present (for example when testing the HTML links
in `gitlab-org/gitlab`), Nanoc will fail when
building the site. By skipping the addition of
the feature flags content in `ee/user/feature_flags.md`,
Nanoc can build the site successfully.