Refactor edit_on_gitlab helper
Rather than having a bunch of case statements with URLs, this changes
the `edit_on_gitlab` helper so that it builds urls from a project path,
a branch name, and a docs file path. This is easier to maintain since
it's a data-driven approach. The PRODUCT_REPOS data can be edited
without changing the program logic.